[ebpf] Separate DNS files
