+++ pkg/ebpf/dns_snooper_test.go
@@ -1,3 +1,5 @@
+// +build linux_bpf
+
 package ebpf
 
 import (
